src/sample_functions/write-to-bucket/requirements.txt
google-cloud-storage>=2.8.0
flask>=2.0.0
